requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
SpeechRecognition
pyaudio
google-cloud-speech
Flask-Mail
celery
redis